Utils: Rename `utils.py` to `misc.py` to not confuse PyCharm checker
