+++ rare/components/dialogs/uninstall_dialog.py
@@ -11,7 +11,7 @@ from PyQt5.QtWidgets import (
 )
 
 from legendary.models.game import Game
-from rare.utils.utils import icon
+from rare.utils.misc import icon
 
 
 class UninstallDialog(QDialog):
