+++ rare/components/tabs/downloads/download_thread.py
@@ -13,7 +13,7 @@ from legendary.models.downloading import WriterTask
 
 from rare.shared import GlobalSignalsSingleton
 from rare.models.install import InstallQueueItemModel
-from rare.utils.utils import create_desktop_link
+from rare.utils.misc import create_desktop_link
 from rare.lgndr.downloading import UIUpdate
 
 logger = getLogger("DownloadThread")
