+++ rare/components/tabs/games/game_info/game_settings.py
@@ -10,7 +10,7 @@ from rare.components.tabs.settings import DefaultGameSettings
 from rare.components.tabs.settings.widgets.pre_launch import PreLaunchSettings
 from rare.utils import config_helper
 from rare.utils.extra_widgets import PathEdit
-from rare.utils.utils import icon, WineResolver, get_raw_save_path
+from rare.utils.misc import icon, WineResolver, get_raw_save_path
 
 logger = getLogger("GameSettings")
