+++ rare/components/tabs/settings/widgets/wrapper.py
@@ -11,7 +11,7 @@ from PyQt5.QtWidgets import QHBoxLayout, QLabel, QPushButton, QInputDialog, QFra
 from rare import shared
 from rare.ui.components.tabs.settings.widgets.wrapper import Ui_WrapperSettings
 from rare.utils import config_helper
-from rare.utils.utils import icon
+from rare.utils.misc import icon
 
 logger = getLogger("Wrapper Settings")
