+++ rare/components/tabs/shop/game_info.py
@@ -17,7 +17,7 @@ from rare.shared import LegendaryCoreSingleton
 from rare.components.tabs.shop.shop_models import ShopGame
 from rare.ui.components.tabs.store.shop_game_info import Ui_shop_info
 from rare.utils.extra_widgets import WaitingSpinner, ImageLabel
-from rare.utils.utils import icon
+from rare.utils.misc import icon
 
 logger = logging.getLogger("ShopInfo")
