+++ rare/utils/misc.py
